Add Bond model
Add a Bond model with all required fields and validators for each field.
Add a set containing all valid currency codes to utils.py for validation of
of the currency_code field. The Bond model has a unique_together contraint
on the user and isin fields, as the isin uniquely identifies a bond for each
user. Also add an index on the user and isin column as the lookup on user and
isin from the GET bond view will happen often.