Add endpoint to enable a user to delete a stored bond
Also, because I've decided to make bonds read only after creation, a user
who wishes to modify a bond they have saved will first need to delete the
bond. Tests have been added to cover this. Also, the views for listing,
creating, and deleting have been grouped into a ViewSet, and a SimpleRouter
is now used to create the urlpatterns, which required some changes to the
tests. Also fix a bug in tests where the checking of the stored bonds
after operations was not done on a per-user basis.